Geocode Addresses with the AZM Geocoding Service

This function interacts with the AZM geocoding service to geocode addresses in a data frame, handling batching and type selection.

Usage

azm_geocode(df, address, type = "position", .test_batch = FALSE)

Arguments

df

A data frame containing the addresses to be geocoded.

address

A string indicating the name of the column in df containing the addresses to be geocoded.

type

A string indicating the type of geocoding to be performed. Must be one of "position", "address", or "matchConfidence".

.test_batch

A logical. If TRUE, the function will run in test mode, processing only a small subset of the data.

Value

A data frame containing the original data plus geocoding results.

Examples

azm_geocode(cand, address)
#> Error in get_azm_url(direction = direction): Azure Maps API Subscription Key not found in your environment variables. 
#> Obtain a key from https://azure.com/maps and set it using: 
#> Sys.setenv(AZ_MAPS_SUBSCRIPTION_KEY = 'your_key_here')